Move slf module to new file_formats module
