+++ rust/src/file_formats/mod.rs
@@ -1,0 +1,3 @@
+//! This module contains code to read and write fileformats that are used by Jagged Alliance 2
+
+pub mod slf;
